# Optimize match via short-circuiting