Once we got an empty node set, it's useless to traverse the rest of string, so we return immediately.

I got a 17% speedup:

```
$ hyperfine --warmup 10 "./array-bool 'lean|rust' /usr/share/dict/american-english"  "./array-bool-shortcircuit 'lean|rust' /usr/share/dict/american-english"
Benchmark 1: ./array-bool 'lean|rust' /usr/share/dict/american-english
  Time (mean ± σ):     265.5 ms ±   7.1 ms    [User: 260.0 ms, System: 5.5 ms]
  Range (min … max):   258.5 ms … 280.0 ms    11 runs
 
Benchmark 2: ./array-bool-shortcircuit 'lean|rust' /usr/share/dict/american-english
  Time (mean ± σ):     226.3 ms ±   5.9 ms    [User: 218.5 ms, System: 7.7 ms]
  Range (min … max):   220.7 ms … 240.0 ms    13 runs
 
Summary
  ./array-bool-shortcircuit 'lean|rust' /usr/share/dict/american-english ran
    1.17 ± 0.04 times faster than ./array-bool 'lean|rust' /usr/share/dict/american-english
```